Issue #3109: Guard against no option name found  (3.0)

+++ components/Migrate-CPTUI/Migrate-CPTUI.php
@@ -48,19 +48,24 @@ class Pods_Migrate_CPTUI extends Pods_Component {
 
 	private $taxonomy_option_name = null;
 
-	private $post_types = null;
+	private $post_types = array();
 
-	private $taxonomies = null;
+	private $taxonomies = array();
 
 	/**
 	 * {@inheritdoc}
 	 */
 	public function __construct() {
 
-		$this->post_option_name     = $this->get_option_name( $this->post_option_name_list );
+		$this->post_option_name = $this->get_option_name( $this->post_option_name_list );
+		if ( ! isnull( $this->post_option_name ) ) {
+			$this->post_types = (array) get_option( $this->post_option_name, array() );
+		}
+
 		$this->taxonomy_option_name = $this->get_option_name( $this->taxonomy_option_name_list );
-		$this->post_types           = (array) get_option( $this->post_option_name, array() );
-		$this->taxonomies           = (array) get_option( $this->taxonomy_option_name, array() );
+		if ( ! isnull( $this->taxonomy_option_name ) ) {
+			$this->taxonomies = (array) get_option( $this->taxonomy_option_name, array() );
+		}
 
 	}
 
@@ -148,15 +153,23 @@ class Pods_Migrate_CPTUI extends Pods_Component {
 
 		if ( 1 == pods_v( 'cleanup', $params, 0 ) ) {
 			if ( ! empty( $post_types ) ) {
-				update_option( $this->post_option_name, $post_types );
+				if ( ! isnull( $this->post_option_name ) ) {
+					update_option( $this->post_option_name, $post_types );
+				}
 			} else {
-				delete_option( $this->post_option_name );
+				if ( ! isnull( $this->post_option_name ) ) {
+					delete_option( $this->post_option_name );
+				}
 			}
 
 			if ( ! empty( $taxonomies ) ) {
-				update_option( $this->taxonomy_option_name, $taxonomies );
+				if ( ! isnull( $this->taxonomy_option_name ) ) {
+					update_option( $this->taxonomy_option_name, $taxonomies );
+				}
 			} else {
-				delete_option( $this->taxonomy_option_name );
+				if ( ! isnull( $this->taxonomy_option_name ) ) {
+					delete_option( $this->taxonomy_option_name );
+				}
 			}
 		}
 
@@ -330,8 +343,13 @@ class Pods_Migrate_CPTUI extends Pods_Component {
 	 */
 	public function clean() {
 
-		delete_option( $this->post_option_name );
-		delete_option( $this->taxonomy_option_name );
+		if ( ! isnull( $this->post_option_name ) ) {
+			delete_option( $this->post_option_name );
+		}
+
+		if ( ! isnull( $this->taxonomy_option_name ) ) {
+			delete_option( $this->taxonomy_option_name );
+		}
 
 	}
